fix: align miserable text on the front
The text on the front of the card came close to the card boundary
(i.e. near the bleed or edge), risking cutoff.
This update fixes a line that uses the va keyword consistently
in the text autofit function and changes the font size.
Moreover, it takes into account the bleed, if specified.
Room is also made for padding, if required.